Refactor server code
- Split api files up into logical modules
  - Use [routes -> cotroller -> model + actions] flow
- Added dockerfiles and docker compose scripts
  - Including mongodb for local dev
  - Live reload working on dockerfiles
  - Still need to add configs for prod deploy
- seporate bot and api logic
- Create CoC module to share logic between api and bot
- removed un-needed commands from bot
