+++ bot/.env.sample
@@ -7,5 +7,4 @@ EMAIL=<CoC email for bot>
 PASSWORD=<CoC bot>
 ADMIN_USER=admin
 ADMIN_PASS=password
-DB_USER=<mongodb cloud username>
-DB_PASS=<mongodb cloud password>
+MONGODB_URI=<URI to mongo instance, if in cloud: mongodb+srv://USER:PASSWORD@cluster0.lduex.mongodb.net/bot?retryWrites=true&w=majority>
