+++ bot/models/db.js
@@ -3,9 +3,11 @@ let client = null;
 let matchInfo = null;
 let commandsInfo = null;
 let db = null;
+
 const init = async () => {
-    const uri = `mongodb+srv://${process.env.DB_USER}:${process.env.DB_PASS}@cluster0.lduex.mongodb.net/bot?retryWrites=true&w=majority`;
+    const uri = process.env.MONGODB_URI
     client = new MongoClient(uri, { useUnifiedTopology: true });
+
     await client.connect();
     db = client.db("bot");
     matchInfo = db.collection("matches");
@@ -44,6 +46,7 @@ const addMatch = async (channelName, matchId, removeCurrent = false) => {
     });
     return op;
 };
+
 const removeCurrentMatch = async (channelName) => {
     const op = await matchInfo.updateOne(
         { _id: channelName },
@@ -51,53 +54,60 @@ const removeCurrentMatch = async (channelName) => {
     );
     return op;
 };
-const getChannelMatches = async (channelName) => {
+
+const getLatestChannelMatch = async (channelName) => {
     const op = await matchInfo.findOne({ _id: channelName });
     return op;
 };
 
-const getChannelCommands = async (channelName) => {
-    const op = await commandsInfo.findOne({ _id: channelName });
+const getChannelMatches = async (channelName) => {
+    const op = await matchInfo.find({ _id: channelName });
     return op;
 };
 
-const addCommand = async (channelName, command, response) => {
-    const op = await commandsInfo.updateOne(
-        { _id: channelName },
-        { $set: { [command]: response } },
-        { upsert: true }
-    );
-    return op;
-};
-const getCommands = async () => {
-    const cursor = commandsInfo.find({});
-    if ((await cursor.count()) === 0) {
-        console.log("No documents found!");
-    }
-    const op = {};
-    await cursor.forEach((m) => {
-        const channelName = m._id;
-        delete m._id;
-        op[channelName] = { ...m };
-    });
-    return op;
-};
-
-const removeCommand = async (channelName, command) => {
-    const op = await commandsInfo.updateOne(
-        { _id: channelName },
-        { $unset: { [command]: "" } }
-    );
-};
+// const getChannelCommands = async (channelName) => {
+//     const op = await commandsInfo.findOne({ _id: channelName });
+//     return op;
+// };
+//
+// const addCommand = async (channelName, command, response) => {
+//     const op = await commandsInfo.updateOne(
+//         { _id: channelName },
+//         { $set: { [command]: response } },
+//         { upsert: true }
+//     );
+//     return op;
+// };
+// const getCommands = async () => {
+//     const cursor = commandsInfo.find({});
+//     if ((await cursor.count()) === 0) {
+//         console.log("No documents found!");
+//     }
+//     const op = {};
+//     await cursor.forEach((m) => {
+//         const channelName = m._id;
+//         delete m._id;
+//         op[channelName] = { ...m };
+//     });
+//     return op;
+// };
+//
+// const removeCommand = async (channelName, command) => {
+//     const op = await commandsInfo.updateOne(
+//         { _id: channelName },
+//         { $unset: { [command]: "" } }
+//     );
+// };
 
 module.exports = {
     init,
     getAll,
     addMatch,
     removeCurrentMatch,
+    getLatestChannelMatch,
     getChannelMatches,
-    getChannelCommands,
-    addCommand,
-    getCommands,
-    removeCommand,
+    // getChannelCommands,
+    // addCommand,
+    // getCommands,
+    // removeCommand,
 };
